Switch from zsh to fish
I've tried fish for a couple of weeks now and really like it.
Out of the box it comes with most things one needs, and things I didn't
know I was missing in zsh.

I've `omf` to install the `sushi` theme. And am pretty happy with it.
Only downside is that "bashisms" do not work, but it's possible to work
around that.